Feat: Connection to Database with TypeORM
It is possible to manage all three parts
- GraphQL(TypeDefs), TypeORM(Database Stucture) and DTO(Data Transfer Object) -
by creating one Entity file.
It is possible to extend the Entities and other DTOs
by using OmitType and PartialType from @nestjs/graphql
which actually let the DTOs to be managed programmatically
because by extending the entity, the connection between them has made.