Feat: User Module
Create User Module and following sub parts: resolver, service, entity and dtos.
Pay attention to 'user.entity.ts' file.
You can add methods to check or modify the data before practically insert into DB
and also you can make a method which checks the password's availability
because it's also the typescript's class which is instanciated in 'users.service.ts'.
As you can see in the file mentioned above,
in login method, 'const user = await this.users.findOne...' returns the
User instance which has checkPassword method which defined in 'user.entity.ts'.
Note that users in 'this.users.findOne()' and user which returned from 'this.users.findOne()'
are different.
The former is relevant to TypeORM and the latter is to Typescript.